requirements.txt
pygame # Base for the game UI
matplotlib
numpy
neat-python # For the AI